docs: migrate documentation
still WIP, missing code generation (although I'm not sure we want it to
begin with), and hence the events and actions pages (v6 has simplified
both so the need to have them now is not really there imo).

Also missing the "available libraries" page, but I think this shouldn't
have been there to begin with, because I don't want extensions to depend
on Ulaunchers deps, so the extensions break when we refactor away deps.

Main page is still not in place either. I added the title and
description only.